requirements.txt
crewai
duckduckgo-search
streamlit
google-generativeai
python-dotenv